refactor: rename loader file name
