+++ verified-reviews/manifest.gen.ts
@@ -5,14 +5,14 @@
 import * as $$$0 from "./loaders/productDetailsPage.ts";
 import * as $$$1 from "./loaders/productList.ts";
 import * as $$$2 from "./loaders/productListingPage.ts";
-import * as $$$3 from "./loaders/storeReviews.ts";
+import * as $$$3 from "./loaders/storeReview.ts";
 
 const manifest = {
   "loaders": {
     "verified-reviews/loaders/productDetailsPage.ts": $$$0,
     "verified-reviews/loaders/productList.ts": $$$1,
     "verified-reviews/loaders/productListingPage.ts": $$$2,
-    "verified-reviews/loaders/storeReviews.ts": $$$3,
+    "verified-reviews/loaders/storeReview.ts": $$$3,
   },
   "name": "verified-reviews",
   "baseUrl": import.meta.url,
